web-animations/interfaces/AnimationEffectTiming/delay.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationEffectTiming.delay</title>
<link rel="help" href="https://drafts.csswg.org/web-animations/#dom-animationeffecttiming-delay">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<script src="../../resources/timing-tests.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(t => {
  const anim = createDiv(t).animate(null, 100);
  anim.effect.timing.delay = 100;
  assert_equals(anim.effect.timing.delay, 100, 'set delay 100');
  assert_equals(anim.effect.getComputedTiming().delay, 100,
                  'getComputedTiming() after set delay 100');
}, 'Allows setting the delay to a positive number');

test(t => {
  const anim = createDiv(t).animate(null, 100);
  anim.effect.timing.delay = -100;
  assert_equals(anim.effect.timing.delay, -100, 'set delay -100');
  assert_equals(anim.effect.getComputedTiming().delay, -100,
                'getComputedTiming() after set delay -100');
}, 'Allows setting the delay to a negative number');

test(t => {
  const anim = createDiv(t).animate(null, 100);
  anim.effect.timing.delay = 100;
  assert_equals(anim.effect.getComputedTiming().progress, null);
  assert_equals(anim.effect.getComputedTiming().currentIteration, null);
}, 'Allows setting the delay of an animation in progress: positive delay that'
   + ' causes the animation to be no longer in-effect');

test(t => {
  const anim = createDiv(t).animate(null, { fill: 'both', duration: 100 });
  anim.effect.timing.delay = -50;
  assert_equals(anim.effect.getComputedTiming().progress, 0.5);
}, 'Allows setting the delay of an animation in progress: negative delay that'
   + ' seeks into the active interval');

test(t => {
  const anim = createDiv(t).animate(null, { fill: 'both', duration: 100 });
  anim.effect.timing.delay = -100;
  assert_equals(anim.effect.getComputedTiming().progress, 1);
  assert_equals(anim.effect.getComputedTiming().currentIteration, 0);
}, 'Allows setting the delay of an animation in progress: large negative delay'
   + ' that causes the animation to be finished');

for (const invalid of gBadDelayValues) {
  test(t => {
    const anim = createDiv(t).animate(null);
    assert_throws({ name: 'TypeError' }, () => {
      anim.effect.timing.delay = invalid;
    });
  }, `Throws when setting invalid delay value: ${invalid}`);
}

</script>
</body>
